Made the ANN
The ANN will work on any of the given dataset, given we clean and extract the training and testing data from the given one. We need to look into the overfitting which is looked into by p=0.1 using Dropout class in keras. The Dataset is large, so we need to run this python script in the cloud (Azure here).